some fixed + tests, added README.md
