+++ .gitignore
@@ -5,8 +5,7 @@
 /ms-api/target/
 /e2e-tests/target/
 /ms-consumer/target/
-
-/journal-service/src/resources/application.properties
+/journal-service/target/
 
 #IDEA
 *.iml
