+++ ms-api/src/main/java/ua/ms/configuration/security/AuthManager.java
@@ -1,7 +1,7 @@
 package ua.ms.configuration.security;
 
-import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -10,18 +10,24 @@ import org.springframework.security.core.AuthenticationException;
 import org.springframework.stereotype.Component;
 import ua.ms.configuration.security.service.RegistrationService;
 import ua.ms.entity.user.User;
-import ua.ms.util.journal.EventServiceWrapper;
+import ua.ms.util.journal.EventServiceImpl;
 
 import java.util.Collections;
 import java.util.Optional;
 
 @Log4j2
 @Component
-@RequiredArgsConstructor
 public class AuthManager implements AuthenticationManager {
 
     private final RegistrationService registrationService;
-    private final EventServiceWrapper eventJournalService;
+
+    private final EventServiceImpl eventJournalService;
+
+    public AuthManager(@Qualifier("registrationService") RegistrationService registrationService,
+                       @Qualifier("eventJournalService") EventServiceImpl eventJournalService) {
+        this.registrationService = registrationService;
+        this.eventJournalService = eventJournalService;
+    }
 
     @Override
     public Authentication authenticate(Authentication authentication) throws AuthenticationException {
